Cat generation fully functional (?)
Cats now only spawn when there is both food and water.  The app now updates the cat list every time a new cat is generated (though you have to close and reopen the cat list to see them, we could maybe add the ability to refresh the list without closing it?  If we have time.)